pyproject.toml
[project]
name = "spot-booker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "python-dotenv>=1.1.1",
    "selenium-driverless>=1.9.4",
]
